Move public AG reconciler to `activegate` package
