+++ src/controllers/dynakube/activegate/activegate/config.go
@@ -1,0 +1,5 @@
+package activegate
+
+import "github.com/Dynatrace/dynatrace-operator/src/logger"
+
+var log = logger.NewDTLogger().WithName("activegate-reconciler")
