+++ src/controllers/dynakube/activegate/activegate/reconciler.go
@@ -1,4 +1,4 @@
-package commonReconciler
+package activegate
 
 import (
 	"context"
