--- src/controllers/dynakube/activegate/commonReconciler/config.go
@@ -1,5 +1,0 @@
-package commonReconciler
-
-import "github.com/Dynatrace/dynatrace-operator/src/logger"
-
-var log = logger.NewDTLogger().WithName("activegate-commonReconciler")
